backend/requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-CORS
Flask-PyMongo
